+++ modules/api/jest.config.cjs
@@ -1,0 +1,10 @@
+const config = {
+    verbose: true,
+    roots: [
+        'src/main/js/',
+        'src/test/js/'
+    ],
+    transform: {}
+};
+
+module.exports = config;