+++ modules/api/src/main/js/lib/HttpError.js
@@ -1,0 +1,7 @@
+export default class HttpError extends Error {
+    constructor(status, originalUrl) {
+        super();
+        this.status = status;
+        this.originalUrl = originalUrl;
+    }
+}